governor print

+++ studytool/st_idle_freq.c
@@ -13,7 +13,7 @@
 // https://docs.kernel.org/admin-guide/pm/cpuidle.html
 
 
-// get functions
+// get functions. Convert last char to \0 to avoid newline
 int read_string_addr(const char * addr, char * buffer, size_t buffer_len)
 {
     int fd = open(addr, O_RDONLY);
@@ -117,7 +117,7 @@ PackageStats st_get_package()
     read_string_addr(PACKAGE_CURRENT_GOVERNOR_ADDR, package.current_governor, sizeof(package.current_governor));
     
     read_string_addr(PACKAGE_AVAILABLE_GOVERNORS_ADDR, package.available_governors, sizeof(package.available_governors));
-    for ( int i = 0; i < strlen(package.available_governors); ++i) // convert to multi-string / string table format
+    for (size_t i = 0; i < strlen(package.available_governors); ++i) // convert to multi-string / string table format
     {
         if(package.available_governors[i] == '\n')
         {
